+++ src/lib/models/book.rs
@@ -6,12 +6,11 @@ use rusqlite::Row;
 use serde::Serialize;
 
 use crate::applebooks::database::{ABDatabaseName, ABQuery};
-use crate::utils;
 
 use super::datetime::DateTimeUtc;
 
 /// A struct represening a book and its metadata.
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default, Clone, Serialize)]
 pub struct Book {
     /// The title of the book.
     pub title: String,
@@ -26,20 +25,6 @@ pub struct Book {
 
     /// The book's metadata.
     pub metadata: BookMetadata,
-}
-
-impl Book {
-    ///Returns a slugified string of the title.
-    #[must_use]
-    pub fn slug_title(&self) -> String {
-        utils::to_slug_string(&self.title, '-')
-    }
-
-    ///Returns a slugified string of the author.
-    #[must_use]
-    pub fn slug_author(&self) -> String {
-        utils::to_slug_string(&self.author, '-')
-    }
 }
 
 impl ABQuery for Book {
@@ -67,41 +52,6 @@ impl ABQuery for Book {
                 last_opened: last_opened.into(),
             },
         }
-    }
-}
-
-impl serde::Serialize for Book {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        #[derive(Serialize)]
-        struct _Slugs {
-            title: String,
-            author: String,
-        }
-
-        #[derive(Serialize)]
-        struct _Book<'a> {
-            title: &'a str,
-            author: &'a str,
-            tags: &'a BTreeSet<String>,
-            metadata: &'a BookMetadata,
-            slugs: _Slugs,
-        }
-
-        let book = _Book {
-            title: &self.title,
-            author: &self.author,
-            tags: &self.tags,
-            metadata: &self.metadata,
-            slugs: _Slugs {
-                title: self.slug_title(),
-                author: self.slug_author(),
-            },
-        };
-
-        book.serialize(serializer)
     }
 }